Remove expected NumberFormatter class. Now just the doubleToString method expected and implemented.
