+++ src/darwinMain/kotlin/com/configcat/NumberFormatter.kt
@@ -1,17 +1,16 @@
 package com.configcat
 
-public class IosNumberFormatter : NumberFormatter {
-    override fun doubleToString(doubleToString: Double): String {
-        return doubleToString.toString()
-        // Handle Double.NaN, Double.POSITIVE_INFINITY and Double.NEGATIVE_INFINITY
+internal actual fun doubleToString(doubleToString: Double): String {
+    return doubleToString.toString()
+    // Handle Double.NaN, Double.POSITIVE_INFINITY and Double.NEGATIVE_INFINITY
 //        if (doubleToString.isNaN() || doubleToString.isInfinite()) {
 //            return doubleToString.toString()
 //        }
 
-        // To get similar result between different SDKs the Double value format is modified.
-        // Between 1e-6 and 1e21 we don't use scientific-notation. Over these limits scientific-notation used but the
-        // ExponentSeparator replaced with "e" and "e+".
-        // "." used as decimal separator in all cases.
+    // To get similar result between different SDKs the Double value format is modified.
+    // Between 1e-6 and 1e21 we don't use scientific-notation. Over these limits scientific-notation used but the
+    // ExponentSeparator replaced with "e" and "e+".
+    // "." used as decimal separator in all cases.
 //        val abs = abs(doubleToString)
 //        val formatter = NSNumberFormatter()
 //        formatter.minimumFractionDigits = 0u
@@ -28,9 +27,4 @@ public class IosNumberFormatter : NumberFormatter {
 //        }
 //        fmt.decimalFormatSymbols = symbols
 //        return fmt.format(doubleToString)
-    }
-}
-
-internal actual fun numberFormatter(): NumberFormatter {
-    return IosNumberFormatter()
 }
