+++ src/nativeMain/kotlin/com/configcat/NumberFormatter.kt
@@ -1,6 +1,5 @@
 package com.configcat
 
-internal actual fun numberFormatter(): NumberFormatter {
-    // TODO is this valid? what is based on the native?
-    return DefaultNumberFormatter()
+internal actual fun doubleToString(doubleToString: Double): String {
+    return commonDoubleToString(doubleToString)
 }
